requirements.txt
grpcio
grpcio-tools
protobuf
matplotlib